scales of electromagnetic spectrum
for reference. it is surprisingly difficult to achieve the desired effect using separate axes instances in matplotlib.
anyway it doesn't allow for zooming because the x-axis isn't shared. it may be possible to share the x-axis and change the labels,
but there is desired offset of the ticks to correspond to powers of 10