added shape-plots and fixed van't hoff
these normalize the experimental variables and look at the shape of the
resulting isotherm or other data. this can allow comparisons between
experiments with significantly different (though in no way does a
quantitative corresponding states apply).

commit 448475 suggested the Van't hoff plots are unphysical, but in fact
the data shows the adsorption is increasing with 1/T, which is
decreasing with temperature, as expected. the signs have been corrected
and it is shown that both the enthalpy of adsorption and entropy of
adsorption tend to be negative, as is physically expected. adsorption
releases energy but confines a molecule and so lowers the number of
complexions